refactor: move API Access to Advanced tab
