+++ frontend/components.d.ts
@@ -14,6 +14,7 @@ declare module 'vue' {
     AddGroupMembersModal: typeof import('./src/components/Modals/AddGroupMembersModal.vue')['default']
     AddGroupModal: typeof import('./src/components/Modals/AddGroupModal.vue')['default']
     AddMemberModal: typeof import('./src/components/Modals/AddMemberModal.vue')['default']
+    AdvancedSettings: typeof import('./src/components/Settings/AdvancedSettings.vue')['default']
     AppSidebar: typeof import('./src/components/AppSidebar.vue')['default']
     AppsMenu: typeof import('./src/components/AppsMenu.vue')['default']
     AttachmentCapsule: typeof import('./src/components/AttachmentCapsule.vue')['default']
